# internal/get/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "404 Downloader API",
        "version": "1.0.0",
        "description": "Local HTTP API used by the in-process 404 Downloader web UI. This API is intended for the local browser session started by 4dl, not as a public external service contract."
    },
    "servers": [
        {
            "url": "/",
            "description": "Current local web server"
        }
    ],
    "paths": {
...
        "/api/search": {
            "post": {
                "operationId": "searchTasks",
                "summary": "Search downloadable tasks",
                "description": "Searches configured providers, stores the returned candidates as the current search result list, and emits an SSE state update. This does not create download tasks and does not add torrents to anacrolix/torrent.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Search results sorted by seeders descending. Results without hash are dropped.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SearchResult"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "502": {
                        "$ref": "#/components/responses/ProviderError"
                    },
                    "503": {
                        "$ref": "#/components/responses/ServiceUnavailable"
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "operationId": "listTasks",
                "summary": "List tasks",
                "description": "Returns the current task list snapshot. List items usually omit file details to keep the response small.",
                "responses": {
                    "200": {
                        "description": "Current application state.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppState"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "createTask",
                "summary": "Create download",
                "description": "Creates a persisted local download task from a selected search result. This is the first point where a search candidate becomes managed download state; result.hash is required and is the only dedupe key. The request path is a per-task download directory; when omitted, the server default downloadDir is used. Tasks and fetched metainfo are stored under --state-dir, or an independent OS config application directory when --state-dir is omitted, and restored on restart.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Download task created, or an existing task with the same hash returned.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskState"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    }
                }
            }
        },
...
        "/api/tasks/{id}": {
            "get": {
                "operationId": "getTask",
                "summary": "Get task",
                "description": "Returns one task with full state. File details are included when metadata is ready.",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TaskID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task state.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskState"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/TaskNotFound"
                    }
                }
            }
        },
        "/api/tasks/{id}/start": {
            "post": {
                "operationId": "startTask",
                "summary": "Start task download",
                "description": "Starts or resumes the task. This is a long-running control action: the current implementation returns 202 Accepted on success, but may wait for torrent metadata before sending that response. Automation and AI agents should not treat the 202 response as completion of the download. Use a client-side timeout appropriate for metadata lookup, then observe progress through GET /api/tasks/{id} or the /api/tasks/stream SSE endpoint.",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TaskID"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/EmptyJSON"
                },
                "responses": {
                    "202": {
                        "description": "Download started or resumed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskState"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/TaskNotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    }
                }
            }
        },
        "/api/tasks/{id}/pause": {
            "post": {
                "operationId": "pauseTask",
                "summary": "Pause task download",
                "description": "Pauses an active task download. If the task has no active runtime, the API returns the current item state.",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TaskID"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/EmptyJSON"
                },
                "responses": {
                    "200": {
                        "description": "Download paused or current state returned.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskState"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/TaskNotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    }
                }
            }
        },
        "/api/tasks/{id}/delete": {
            "post": {
                "operationId": "deleteTask",
                "summary": "Delete downloaded data",
                "description": "Cancels the task runtime, removes the persisted task, and attempts to delete final payload files and their .part files.",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/TaskID"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/EmptyJSON"
                },
                "responses": {
                    "200": {
                        "description": "Local download state deleted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskState"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/TaskNotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    }
                }
            }
        }
    },
    "components": {
        "parameters": {
            "TaskID": {
                "name": "id",
                "in": "path",
                "required": true,
                "description": "Task item identifier.",
                "schema": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "requestBodies": {
            "EmptyJSON": {
                "description": "Request body is currently ignored. The web UI sends an empty JSON object.",
                "required": false,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        },
                        "example": {}
                    }
                }
            }
        },
        "responses": {
            "BadRequest": {
                "description": "The request was invalid.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        }
                    }
                }
            },
            "TaskNotFound": {
                "description": "Task was not found.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        },
                        "example": {
                            "error": "task not found"
                        }
                    }
                }
            },
            "Conflict": {
                "description": "The requested task operation could not be completed.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        }
                    }
                }
            },
            "ProviderError": {
                "description": "One or more search providers failed.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        }
                    }
                }
            },
            "ServiceUnavailable": {
                "description": "Search is not configured.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/APIError"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "SearchRequest": {
                "type": "object",
                "required": [
                    "query"
                ],
                "properties": {
                    "query": {
                        "type": "string",
                        "minLength": 1
                    },
                    "limit": {
                        "type": "integer",
                        "minimum": 1,
                        "description": "Maximum number of results to return. Defaults to the server limit when omitted or zero."
                    },
                    "providers": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "Provider names to query. Omit or pass an empty array to search all configured providers."
                    }
                },
                "additionalProperties": false
            },
            "CreateTaskRequest": {
                "type": "object",
                "required": [
                    "result"
                ],
                "properties": {
                    "result": {
                        "$ref": "#/components/schemas/SearchResult"
                    },
                    "path": {
                        "type": "string",
                        "description": "Per-task download directory. Omit or pass an empty string to use the server default downloadDir."
                    }
                },
                "additionalProperties": false
            },
            "SearchResult": {
                "type": "object",
                "required": [
                    "provider",
                    "title",
                    "hash",
                    "seeders",
                    "peers"
                ],
                "properties": {
                    "provider": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "bytes": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "category": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string"
                    },
                    "seeders": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "peers": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "hash": {
                        "type": "string",
                        "description": "BitTorrent info hash. Search results without hash are not returned."
                    },
                    "magnetUrl": {
                        "type": "string"
                    }
                },
                "additionalProperties": false
            },
            "AppState": {
                "type": "object",
                "required": [
                    "updated",
                    "downloadDir",
                    "searchResults",
                    "tasks"
                ],
                "properties": {
                    "updated": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Time when this snapshot was generated."
                    },
                    "downloadDir": {
                        "type": "string",
                        "description": "Default download directory used when a task does not specify path."
                    },
                    "searchResults": {
                        "type": "array",
                        "description": "Current search candidates. These are not download tasks and are not added to anacrolix/torrent until selected for download.",
                        "items": {
                            "$ref": "#/components/schemas/SearchResult"
                        }
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TaskState"
                        }
                    }
                },
                "additionalProperties": false
            },
            "TaskState": {
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "provider",
                    "seeders",
                    "peers",
                    "path",
                    "downloading",
                    "download",
                    "runtime"
                ],
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "provider": {
                        "type": "string"
                    },
                    "bytes": {
                        "type": "integer",
                        "format": "int64",
                        "minimum": 0
                    },
                    "category": {
                        "type": "string"
                    },
                    "date": {
                        "type": "string",
                        "description": "Provider-supplied date string."
                    },
                    "seeders": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "peers": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "hash": {
                        "type": "string",
                        "description": "BitTorrent info hash. Required for search results added through POST /api/tasks."
                    },
                    "magnetUrl": {
                        "type": "string",
                        "description": "Magnet link when available."
                    },
                    "downloading": {
                        "type": "boolean"
                    },
                    "download": {
                        "$ref": "#/components/schemas/DownloadView"
                    },
                    "error": {
                        "type": "string"
                    },
                    "files": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FileItem"
                        }
                    },
                    "runtime": {
                        "$ref": "#/components/schemas/RuntimeView"
                    },
                    "path": {
                        "type": "string",
                        "description": "Per-task download directory."
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}